+++ src/index.ts
@@ -6,3 +6,5 @@ export * from "./request-executor";
 export * from "./error-handler";
 export * from "./progress-tracker";
 export * from "./pipeline-orchestrator";
+export * from "./pipeline-builder";
+export * from "./pipeline-validator";